app/requirements.txt
asgiref==3.7.2
Django==5.0
django-jazzmin==2.6.0
sqlparse==0.4.4
djangorestframework==3.14.0
Pillow==10.1.0
